+++ relay-server/src/metrics_extraction/transactions.rs
@@ -5,7 +5,7 @@ use std::collections::BTreeSet;
 use {
     relay_common::UnixTimestamp,
     relay_general::protocol::{AsPair, Event, EventType},
-    relay_general::store::{get_breakdown_measurements, BreakdownsConfig},
+    relay_general::store::{get_breakdown_measurements, normalize_dist, BreakdownsConfig},
     relay_metrics::{Metric, MetricUnit, MetricValue},
     std::collections::BTreeMap,
     std::fmt::Write,
@@ -35,7 +35,7 @@ fn metric_name(parts: &[&str]) -> String {
 }
 
 #[cfg(feature = "processing")]
-fn transaction_status(transaction: &Event) -> Option<String> {
+fn extract_transaction_status(transaction: &Event) -> Option<String> {
     use relay_general::{
         protocol::{Context, ContextInner},
         types::Annotated,
@@ -48,6 +48,13 @@ fn transaction_status(transaction: &Event) -> Option<String> {
     };
     let span_status = trace_context.status.value()?;
     Some(span_status.to_string())
+}
+
+#[cfg(feature = "processing")]
+fn extract_dist(transaction: &Event) -> Option<String> {
+    let mut dist = transaction.dist.0.clone();
+    normalize_dist(&mut dist);
+    dist
 }
 
 #[cfg(feature = "processing")]
@@ -94,6 +101,9 @@ pub fn extract_transaction_metrics(
     let mut tags = BTreeMap::new();
     if let Some(release) = event.release.as_str() {
         tags.insert("release".to_owned(), release.to_owned());
+    }
+    if let Some(dist) = extract_dist(event) {
+        tags.insert("dist".to_owned(), dist);
     }
     if let Some(environment) = event.environment.as_str() {
         tags.insert("environment".to_owned(), environment.to_owned());
@@ -182,7 +192,7 @@ pub fn extract_transaction_metrics(
         unit: MetricUnit::Duration(DurationPrecision::MilliSecond),
         value: MetricValue::Distribution(duration_millis as f64),
         timestamp,
-        tags: match transaction_status(event) {
+        tags: match extract_transaction_status(event) {
             Some(status) => with_tag(&tags, "transaction.status", status),
             None => tags.clone(),
         },
@@ -227,6 +237,7 @@ mod tests {
             "type": "transaction",
             "timestamp": "2021-04-26T08:00:00+0100",
             "release": "1.2.3",
+            "dist": "foo ",
             "environment": "fake_environment",
             "transaction": "mytransaction",
             "tags": {
@@ -313,6 +324,7 @@ mod tests {
         for metric in metrics {
             assert!(matches!(metric.value, MetricValue::Distribution(_)));
             assert_eq!(metric.tags["release"], "1.2.3");
+            assert_eq!(metric.tags["dist"], "foo");
             assert_eq!(metric.tags["environment"], "fake_environment");
             assert_eq!(metric.tags["transaction"], "mytransaction");
             assert_eq!(metric.tags["fOO"], "bar");
@@ -371,6 +383,7 @@ mod tests {
         }
 
         assert_eq!(duration_metric.tags.len(), 4);
+        assert_eq!(duration_metric.tags["release"], "1.2.3");
         assert_eq!(duration_metric.tags["transaction.status"], "ok");
         assert_eq!(duration_metric.tags["environment"], "fake_environment");
         assert_eq!(duration_metric.tags["transaction"], "mytransaction");